Data reformatting as function
I think it is finished now.
Just a few questions:
- what to do with library(...) in the script where you use packages
- what to do with the part where you define all the parts that go into the function? for example number of age classes = 5, summer_removal = c(6,7,8,9)  etc? I left this in the end of the script for now
- I used n_ageclasses as the number of age classes. Then later in the function I also use max_age <- (n_ageclasses -1)

But I later found in another script by Chloe: Amax = number of age classes to be used in the analysis.
Maybe we should mke this the same. But isnt the max age the nr of age classes -1 because they start at age 0?