Add files via upload
This is the script to use microsatellite data to calculate the probabilities that a fox actually comes from Varanger or is an immigrant. The script reformats the genetic data and writes input files for the stand alone program Geneclass 2.0. The analysis itself is run in Geneclass, and the script then reads in the output files and assembles them to a table with a probability for each indivual to originate from the Varanger population.