Tested effect of "brute-forcing" higher harvest mortality
Forcing higher harvest mortality leads to overall lower population size estimates and this works through two mechanisms.
The first is directly through lowering annual survival.
The second is indirect and this works through a decrease in denning survival. Lower denning survival means lower local pup production AND less immigrants (since immigration rate stays the same), which together reduce the number of age class 1 individuals available for harvest.

What this tells me is that it's important we do that sensitivity analysis for the prior on denning survival and - ideally - try to add additional information on it.