Adding summer harvest with toggles
Hei here I added the option of adding the unaged summer to WAaH. and also the option of making a SAaH for specific years.
I could test it all the way through because I am getting an error at gen.data from wrangleData_gen. This is because I don't know where "llotherV2" comes from. Is this something on your computer Chloe?