Changed initialization to start on first "october census"
With the first setup, it was clear that the June population sizes in the first year were non-identifiable parameters. This is likely linked to the fact that summer harvest data is only available from year 2. With this edit, the model now initializes at the October census in year 1 instead of the June census in an attempt to avoid this problem.