Fixed rodentFall 2003, first year is now 2004
I also removed an outdated check that required an equal number of fall and spring seasons in the trapping data. Since we now use fall and spring from different years (winter instead of summer rodent) this is not needed anymore. A later check on line 86 (new script) already checks if the last start_hunting_year for both seasons is the same, if not that year will be NA because it only has one of the seasons

I had to skip the merged_fallwinter file creating because we want fall to start at 2004 and winter at 2003, I didnt see how to do this in the same file.

Is the last line (return(merged_fallwinter )) needed?