Implemented new structure for immigrant number model
Previously, we used a model with discrete uniform distributions for annual immigrant numbers as earlier versions of the IPM did not take well to normalization of immigrant numbers.
Now, this is different and the new model - which uses a lognormal distribution to model time variation in immigrant numbers, works fine.