Ran simplified model with and without genetic data
Trying to uncover where the discrepancy in the simplified model estimating immR vs. simplified model estimating Imm comes from. If genetic data is used, even this first version (with and without age-specific mH, with and without covariates on immigration) already produces the discrepancy. When ignoring genetic data, the discrepancy seems less but chain mixing is poor for some parameters in the models estimating immigrant numbers and it looks like convergence may not be reached yet. Longer runtimes may be necessary.